requirements.txt
streamlit
pandas
python-docx
PyMuPDF
pytesseract
Pillow
matplotlib
seaborn
python-dotenv
together
